[fix] run 실패 문제 해결
- DTO 쪽 임시로 형변환 (RetentionSupport)

+++ momentum-dao-be/src/main/java/com/dao/momentum/retention/prospect/command/domain/aggregate/RetentionSupport.java
@@ -59,9 +59,9 @@ public class RetentionSupport {
                 dto.compLevel(),
                 dto.relationLevel(),
                 dto.growthLevel(),
-                dto.tenureLevel(),
+                BigDecimal.valueOf(dto.tenureLevel()),
                 dto.wlbLevel(),
-                dto.retentionScore(),
+                dto.retentionScore().intValue(),
                 LocalDateTime.now()
         );
     }
